Refactor lib/Engine/Rendering
